requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
passlib[bcrypt]
python-jose[cryptography]
PyMuPDF
Pillow
opencv-python
reportlab
psycopg2-binary
pydantic-settings
pytest
